fix(ram): do not read from cgroup
